Move wearos to the new structure
