+++ wearos/src/main/kotlin/com/boswelja/watchconnection/wearos/Constants.kt
@@ -1,0 +1,5 @@
+package com.boswelja.watchconnection.wearos
+
+object Constants {
+    internal const val WEAROS_PLATFORM = "WEAR_OS"
+}
