+++ wearos/src/main/kotlin/com/boswelja/watchconnection/wearos/WearOSMessageReceiver.kt
@@ -1,7 +1,8 @@
 package com.boswelja.watchconnection.wearos
 
-import com.boswelja.watchconnection.core.Messages
 import com.boswelja.watchconnection.core.Watch
+import com.boswelja.watchconnection.core.message.Messages
+import com.boswelja.watchconnection.wearos.Constants.WEAROS_PLATFORM
 import com.google.android.gms.wearable.MessageEvent
 import com.google.android.gms.wearable.WearableListenerService
 
@@ -13,7 +14,7 @@ class WearOSMessageReceiver : WearableListenerService() {
     override fun onMessageReceived(messageEvent: MessageEvent?) {
         messageEvent?.let { event ->
             // Get watch ID
-            val watchId = Watch.createUUID(WearOSPlatform.PLATFORM, event.sourceNodeId)
+            val watchId = Watch.createUUID(WEAROS_PLATFORM, event.sourceNodeId)
 
             // Send message broadcast
             Messages.sendBroadcast(this, watchId, event.path, event.data)
